fix: REGRESS.FACTORIMP 返回值 long[] → double[] 修复 Excel-DNA 封送失败
根因: Excel-DNA 不支持 Int64 数组 (long[]) 作为 UDF 返回值类型，
Excel C API 无原生 Int64 封送路径，导致 COM 异常 → #VALUE!。

修复:
- RegressionUdf.cs: .Select(i=>(long)i) → .Select(i=>(double)i)
- RegressionUdfTests.cs: (long[]) → (double[]) 类型转换
- api-reference.md: 返回类型 long[] → double[]

double 可精确表示所有 Excel 列索引 (最大 16384 << 2^53)。

已验证: dotnet test 全量 2790+ 通过，CrossVal 157 通过。

+++ src/Analytics/RegressionUdf.cs
@@ -35,7 +35,7 @@ namespace ExcelFormulaLabs.Analytics
 
         [ExcelFunction(Name = "REGRESS.FACTORIMP", Description = "Rank predictor importance by |t| from standardized OLS. Returns 0-based column index array most-to-least important.")]
         public static object UDF_REGRESS_FACTORIMP([ExcelArgument(Name="known_y", Description="The Y variable range (dependent variable)")] object y, [ExcelArgument(Name="known_x", Description="The X variable range (independent variables)")] object X)
-            => OutputWrapper.WrapError(() => RegressionCore.FactorImportance(M(X), V(y)).Select(i=>(long)i).ToArray());
+            => OutputWrapper.WrapError(() => RegressionCore.FactorImportance(M(X), V(y)).Select(i=>(double)i).ToArray());
 
         [ExcelFunction(Name = "REGRESS.COEF", Description = "OLS regression coefficients only (beta vector).")]
         public static object UDF_REGRESS_COEF([ExcelArgument(Name="known_y", Description="The Y variable range (dependent variable)")] object y, [ExcelArgument(Name="known_x", Description="The X variable range (independent variables)")] object X)
